\name{adwave-package}
\alias{adwave-package}
\alias{adwave}
\docType{package}
\title{Wavelet Analysis of Genomic Data from Admixed Populations}
\description{
Implements wavelet-based approaches for describing population admixture. Principal Components Analysis (PCA) is used to define the population structure and produce a localized admixture signal for each individual. Wavelet summaries of the PCA output describe variation present in the data and can be related to population-level demographic processes. For more details, see accompanying paper.
}
\details{
\tabular{ll}{
Package: \tab adwave\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-03\cr
License: GPL (>=2)\cr
}

}
\author{
Jean Sanderson

Maintainer: Murray Cox <m.p.cox@massey.ac.nz>

}
\references{For further details, see accompanying paper.}

\keyword{ package }
\seealso{
\code{\link{signal}},  \code{\link{wavesum}},  \code{\link{plotsignal}}  
}
\examples{
# see examples in individual help files
}
